# api/swagger/swagger.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'API Documentation'
  description: 'This is an API documentation for Explorer'
servers:
  - description: Local Machine server
    url: http://localhost:4000
  - description: Staging Xplorer Server
    url: https://finder.empliance.in
  - description: Poduction Xplorer Server
    url: https://production.finder.empliance.in
...
paths:
...
  /user/get-litigation-bi/details:
    post:
      # security:
      #   - bearerAuth: []
      tags:
        - LITIGATION BI
      summary: Litigation BI Details
      operationId: litigationBiDetails
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                companyId:
                  description: Enter Company CIN
                  type: string       
      responses:
        "200":
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Invalid username/password supplied
...
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT